// src/typings/enums/keyboards.ts
export enum KEYBOARDS {
  PROFILE = "👤 Mening profilim",
  ABOUT_US = "☕️ Fillallarimiz haqida",
  SEND_FEEDBACK = "📝 Fikr-mulohaza yuborish",
  BACK = "🔙 Ortga qaytish",
  STATISTICS = "📊 Statistika",
  FIND_USER = "🔎 Foydalanuvchini topish",
  NEW_MESSAGE = "💌 Yangi xabar",
  ADD_CASHBACK = "💰 Cashback qo'shish",
  REMOVE_CASHBACK = "🤑 Cashbackni yechish",
  UPDATE = "✏️ O'zgartirish",
  SEND_CONTACT = "📲 Kontaktni yuborish",
  BACK_MAIN = "🏠 Bosh menyuga qaytish",
  SKIP = "🔜 O'tkazib yuborish",
  VIDEO = "📹 Video",
  PHOTO = "📸 Rasm",
  YES = "✅ Ha",
  NO = "❌ Yo'q",
}